+++ Software/Development/SemesterProject/App/controller.c
@@ -39,9 +39,9 @@ float PiCalculationID(float reference, float measuredValue)
     return PiCalculation(&PIControllerList.IdController, reference, measuredValue);
 }
 
-float PiCalculationSpeed(float reference, float measuredValue)
+float PiCalculationSpeed(float reference, int16 measuredValue)
 {
-    return PiCalculation(&PIControllerList.SpeedController, reference, measuredValue);
+    return PiCalculation(&PIControllerList.SpeedController, reference, (float)measuredValue);
 }
